django_server/requirements.txt
cloudinary
gunicorn
django==2.0.3
djangorestframework
markdown
django-filter
setuptools
psycopg2-binary
dj-database-url
